requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
pydantic==2.5.0
python-dotenv==1.0.0
websockets==12.0
fastapi-users[redis]==13.0.0
bcrypt==4.1.2
passlib[bcrypt]
python-jose[cryptography]==3.3.0
email-validator